flake8 and small logging fixes

+++ lms/lmsweb/git_service.py
@@ -1,6 +1,6 @@
 import os
 import shutil
-import subprocess
+import subprocess  # noqa: S404
 import tempfile
 import typing
 import logging
@@ -46,29 +46,42 @@ class GitService:
     def handle_operation(self) -> flask.Response:
         git_operation = self._extract_git_operation()
 
-        first_time_repository = not os.path.exists(self.repository_folder)
+        git_repository_folder = os.path.join(self.repository_folder, '.git')
+        first_time_repository = not os.path.exists(git_repository_folder)
         if first_time_repository:
-            os.makedirs(self.repository_folder)
-            p = subprocess.Popen(
-                args=['git', 'init', '--bare', '--initial-branch=main'],
+            os.makedirs(self.repository_folder, exist_ok=True)
+            proc = subprocess.Popen(  # noqa: S603
+                args=['git', 'init', '--bare'],
                 stdin=subprocess.PIPE,
                 stdout=subprocess.PIPE,
                 stderr=subprocess.PIPE,
                 cwd=self.repository_folder,
             )
-            assert p.wait() == 0
+            if proc.wait() != 0:
+                _logger.error(
+                    'Failed to execute command. stdout=%s\nstderr=%s',
+                    proc.stdout.read(), proc.stderr.read(),
+                )
+                raise EnvironmentError
 
         if not git_operation.supported:
             raise EnvironmentError
 
-        p = subprocess.Popen(
+        proc = subprocess.Popen(  # noqa: S603
             args=git_operation.service_command,
             stdin=subprocess.PIPE,
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
             cwd=self._base_repository_folder,
         )
-        data_out, _ = p.communicate(self._request.data, 20)
+        data_out, _ = proc.communicate(self._request.data, 20)
+
+        if proc.wait() != 0:
+            _logger.error(
+                'Failed to execute command. stdout=%s\nstderr=%s',
+                proc.stdout.read(), proc.stderr.read(),
+            )
+            raise EnvironmentError
 
         if git_operation.format_response:
             data_out = git_operation.format_response(data_out)
@@ -87,7 +100,10 @@ class GitService:
         return res
 
     @staticmethod
-    def build_response(data_out: bytes, git_operation: _GitOperation) -> flask.Response:
+    def build_response(
+            data_out: bytes,
+            git_operation: _GitOperation,
+    ) -> flask.Response:
         res = flask.make_response(data_out)
         res.headers['Expires'] = 'Fri, 01 Jan 1980 00:00:00 GMT'
         res.headers['Pragma'] = 'no-cache'
@@ -96,31 +112,49 @@ class GitService:
         return res
 
     def _extract_git_operation(self) -> _GitOperation:
+        stateless_rpc = '--stateless-rpc'
+        advertise_refs = '--advertise-refs'
         upload_pack_command = 'git-upload-pack'
         receive_pack_command = 'git-receive-pack'
+        allowed_commands = [upload_pack_command, receive_pack_command]
+
         supported = True
         format_response = False
         contain_new_commits = False
 
         if self._request.path.endswith(upload_pack_command):
             content_type = 'application/x-git-upload-pack-result'
-            service_command = [upload_pack_command, '--stateless-rpc', self.project_name]
+            service_command = [
+                upload_pack_command,
+                stateless_rpc,
+                self.project_name,
+            ]
 
         elif self._request.path.endswith(receive_pack_command):
             content_type = 'application/x-git-receive-pack-result'
-            service_command = [receive_pack_command, '--stateless-rpc', self.project_name]
+            service_command = [
+                receive_pack_command,
+                stateless_rpc,
+                self.project_name,
+            ]
             contain_new_commits = True
 
         elif self._request.path.endswith('/info/refs'):
             service_name = self._request.args.get('service')
-            service_command = [service_name, '--stateless-rpc', '--advertise-refs', self.project_name]
+            service_command = [
+                service_name,
+                stateless_rpc,
+                advertise_refs,
+                self.project_name,
+            ]
             content_type = f'application/x-{service_name}-advertisement'
-            supported = service_name in [upload_pack_command, receive_pack_command]
+
+            supported = service_name in allowed_commands
 
             def format_response_callback(response_bytes: bytes) -> bytes:
                 packet = f'# service={service_name}\n'
                 length = len(packet) + 4
-                prefix = "{:04x}".format(length & 0xFFFF)
+                prefix = '{:04x}'.format(length & 0xFFFF)
 
                 data = (prefix + packet + '0000').encode()
                 data += response_bytes
@@ -129,7 +163,10 @@ class GitService:
             format_response = format_response_callback
 
         else:
-            _logger.error('Failed to find the git command for route %s', self._request.path)
+            _logger.error(
+                'Failed to find the git command for route %s',
+                self._request.path,
+            )
             raise NotImplementedError
 
         return _GitOperation(
@@ -142,23 +179,32 @@ class GitService:
 
     def _load_files_from_repository(self) -> typing.List[upload.File]:
         with tempfile.TemporaryDirectory() as tempdir:
-            p = subprocess.Popen(
+            proc = subprocess.Popen(  # noqa: S603
                 args=['git', 'clone', self.repository_folder, '.'],
                 stdin=subprocess.PIPE,
                 stdout=subprocess.PIPE,
                 stderr=subprocess.PIPE,
                 cwd=tempdir,
             )
-            return_code = p.wait()
-            assert return_code == 0
+            return_code = proc.wait()
+            if return_code != 0:
+                _logger.error(
+                    'Failed to execute command. stdout=%s\nstderr=%s',
+                    proc.stdout.read(), proc.stderr.read(),
+                )
+                raise EnvironmentError
             to_return = []
             # remove git internal files
             shutil.rmtree(os.path.join(tempdir, '.git'))
-            for root, folders, files in os.walk(tempdir):
+            for root, _, files in os.walk(tempdir):
                 for file in files:
-                    with open(os.path.join(root, file)) as f:
-                        to_return.append(upload.File(
-                            path=os.path.join(os.path.relpath(root, tempdir), file),
-                            code=f.read()
-                        ))
+                    upload_file = self._load_file(file, root, tempdir)
+                    to_return.append(upload_file)
             return to_return
+
+    @staticmethod
+    def _load_file(file_path: str, root: str, tempdir: str) -> upload.File:
+        with open(os.path.join(root, file_path)) as f:
+            file_path = os.path.join(os.path.relpath(root, tempdir), file_path)
+            upload_file = upload.File(path=file_path, code=f.read())
+        return upload_file
